[Chore] DetailView에서 사용한 UIView, UIButton 메서드 Extension 파일로 분리
